Obtain the Module Information from Consensus-Based Investigation

Extract the module information based on the consensus results.

comparePhenotypes(
  C3NAObj_Comparison = C3NAObj1,
  C3NAObj_Reference = C3NAObj2,
  corCutoff = 0.2,
  fdr = 0.05,
  unusefulTaxa = NA,
  verbose = 0,
  nBootstrap = 300
)

Arguments

C3NAObj_Comparison

(Required) The C3NAObj with a single phenotype as Disease.

C3NAObj_Reference

(Required) The C3NAObj with a single phenotype as Control.

corCutoff

(Required) A positive correlation cut-off point, default is 0.2.

fdr

(Required) BH-Adjusted p-value, default is 0.05.

unusefulTaxa

(Optional) A list of taxa names that could not be interpreated biologically, e.g. human_gut, ambiguous_taxa

verbose

(Optinal) Verbose for the module preservation analysis, the higher the more detailed output.

nBootstrap

(Optional) Number of bootstrap for the module preservation.